Rename gFlex package to gflex.
