+++ landlab/components/gflex/flexure.py
